refactor: use EntityTable for WriteOffsView
